pyproject.toml
[project]
name = "ms-fabric-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.3.0",
    "azure-identity>=1.20.0",
    "deltalake>=0.25.4",
    "requests>=2.32.3",
    "cachetools>=5.5.2",
]

[tool.setuptools]
packages = ["helpers", "helpers.clients", "helpers.formatters", "helpers.utils"]

[project.scripts]
mcp = "mcp.cli:app [cli]"